[dropdown-menu] working a11y
