+++ semcore/dropdown-menu/src/index.js
@@ -1,3 +1,1 @@
 export { default } from './DropdownMenu';
-export { default as DDMenu } from './DropdownMenu2';
-export { default as DDMenu3 } from './DropdownMenu3';
